Updated 1.0.1
Set initial Avatar Colors on New Avatars - Some new avatars were starting all white (Default Female and Male). Created a table for default avatar colors, added default data for each avatar, and added initial color assignments to the new avatar selection process.

Avatar Designer Window Adjustments - When editing the avatar, the window was wrapping under the bottom menu and the contents of the window did not fill the lower part of the window. Adjusted the vertical size of the window and set the background to black for the designer window only. Set the background opacity back to .2 when it closes and reopens.

Change Avatar Colors from RGB to Hex - Changed the avatar color database fields to use Hex values instead of RGB saved in the database. Everything now supports diffuse, specular, emissive, and ambient except for the user interface, which will be updated soon. This affected the JSON fetched data and save procedures, local and global avatars, and 3D Internet plugin in development.